+++ test/functional/com/litle/sdk/TestAuthReversal.java
@@ -23,7 +23,7 @@ public class TestAuthReversal {
 	public void simpleAuthReversal() throws Exception{
 		AuthReversal reversal = new AuthReversal();
 		reversal.setLitleTxnId(12345678000L);
-		reversal.setAmount(BigInteger.valueOf(106L));
+		reversal.setAmount(106L);
 		reversal.setPayPalNotes("Notes");
 		
 		AuthReversalResponse response = litle.authReversal(reversal);
