+++ test/functional/com/litle/sdk/TestCapture.java
@@ -24,7 +24,7 @@ public class TestCapture {
 	public void simpleCapture() throws Exception{
 		Capture capture = new Capture();
 		capture.setLitleTxnId(123456000);
-		capture.setAmount(BigInteger.valueOf(106L));
+		capture.setAmount(106L);
 		capture.setPayPalNotes("Notes");
 		
 		CaptureResponse response = litle.capture(capture);
@@ -35,7 +35,7 @@ public class TestCapture {
 	public void simpleCaptureWithPartial() throws Exception{
 		Capture capture = new Capture();
 		capture.setLitleTxnId(123456000);
-		capture.setAmount(BigInteger.valueOf(106L));
+		capture.setAmount(106L);
 		capture.setPartial(true);
 		capture.setPayPalNotes("Notes");
 		
@@ -47,11 +47,11 @@ public class TestCapture {
 	public void complexCapture() throws Exception{
 		Capture capture = new Capture();
 		capture.setLitleTxnId(123456000);
-		capture.setAmount(BigInteger.valueOf(106L));
+		capture.setAmount(106L);
 		capture.setPayPalNotes("Notes");
 		EnhancedData enhancedData = new EnhancedData();
 		enhancedData.setCustomerReference("Litle");
-		enhancedData.setSalesTax(BigInteger.valueOf(50L));
+		enhancedData.setSalesTax(50L);
 		enhancedData.setDeliveryType("TBD");
 		capture.setEnhancedData(enhancedData);
 		capture.setPayPalOrderComplete(true);
